adding finance calculater file

+++ python_introduction/finance_calculator.py
@@ -1,8 +1,8 @@
 
 
 
-monthly_income = int(input("Enter your monthly income: "))
-monthly_expense = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+monthly_expense = float(input("Enter your total monthly expenses: "))
 
 Monthly_savings = monthly_income - monthly_expense
 print("Your monthly savings are ", Monthly_savings)
